Added backend interfacing module
