+++ frontend/src/App.tsx
@@ -1,48 +1,12 @@
-import { useState, useEffect } from 'react'
-import axios from 'axios'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
+import { useState } from 'react'
+import BackendCall from './util/backend'
 import './App.css'
-
-interface MessageResponse {
-  data: {
-    message: string;
-  };
-}
-
-function BackendCall() {
-  const [message, setMessage] = useState<string>(''); // TypeScript enforces the type of the state variable
-
-  useEffect(() => {
-    axios.get('/api/message') // Specify the expected response type
-      .then((response: MessageResponse) => {
-        setMessage(response.data.message); // Update the message state variable with the result
-      })
-      .catch((error: unknown) => {
-        console.error(error);
-      });
-  }, []); // Empty dependency array ensures this runs only once when the component mounts
-
-  return (
-    <div>
-      {message ? message : 'Loading...'} {/* Display a loading message until the API call completes */}
-    </div>
-  );
-}
 
 function App() {
   const [count, setCount] = useState(0)
 
   return (
     <>
-      <div>
-        <a href="https://vite.dev" target="_blank">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
       <h1>Vite + React</h1>
       <div className="card">
         <button onClick={() => setCount((count) => count + 1)}>
